finish tests and testing structure
